requirements.txt
asgiref==3.7.2
django-simple-history==3.4.0
Django==5.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
python-dotenv==1.0.1
whitenoise==6.6.0
wfastcgi==3.0.0
django-cors-headers==4.3.1
pillow==10.4.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-filter==24.3
fcm-django==2.3.1
django-jazzmin==3.0.1
